LayoutTests/fast/js/dfg-arith-add-overflow-check-elimination-predicted-but-not-proven-int-expected.txt

Tests that when values predicted but not proven int are used in a tower of additions, we don't eliminate the overflow check unsoundly.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS bar(1, 2, {f:3}) is 6
PASS bar(1, 2, {f:3}) is 6
PASS bar(1, 2, {f:3}) is 6
PASS bar(1, 2, {f:3}) is 6
PASS bar(1, 2, {f:3}) is 6
PASS bar(1, 2, {f:3}) is 6
PASS bar(1, 2, {f:3}) is 6
PASS bar(1, 2, {f:3}) is 6
PASS bar(1, 2, {f:3}) is 6
PASS bar(1, 2, {f:3}) is 6
PASS bar(1, 2, {f:3}) is 6
PASS bar(1, 2, {f:3}) is 6
PASS bar(1, 2, {f:3}) is 6
PASS bar(1, 2, {f:3}) is 6
PASS bar(1, 2, {f:3}) is 6
PASS bar(1, 2, {f:3}) is 6
PASS bar(1, 2, {f:3}) is 6
PASS bar(1, 2, {f:3}) is 6
PASS bar(1, 2, {f:3}) is 6
PASS bar(1, 2, {f:3}) is 6
PASS bar(1, 2, {f:3}) is 6
PASS bar(1, 2, {f:3}) is 6
PASS bar(1, 2, {f:3}) is 6
PASS bar(1, 2, {f:3}) is 6
PASS bar(1, 2, {f:3}) is 6
PASS bar(1, 2, {f:3}) is 6
PASS bar(1, 2, {f:3}) is 6
PASS bar(1, 2, {f:3}) is 6
PASS bar(1, 2, {f:3}) is 6
PASS bar(1, 2, {f:3}) is 6
PASS bar(1, 2, {f:3}) is 6
PASS bar(1, 2, {f:3}) is 6
PASS bar(1, 2, {f:3}) is 6
PASS bar(1, 2, {f:3}) is 6
PASS bar(1, 2, {f:3}) is 6
PASS bar(1, 2, {f:3}) is 6
PASS bar(1, 2, {f:3}) is 6
PASS bar(1, 2, {f:3}) is 6
PASS bar(1, 2, {f:3}) is 6
PASS bar(1, 2, {f:3}) is 6
PASS bar(1, 2, {f:3}) is 6
PASS bar(1, 2, {f:3}) is 6
PASS bar(1, 2, {f:3}) is 6
PASS bar(1, 2, {f:3}) is 6
PASS bar(1, 2, {f:3}) is 6
PASS bar(1, 2, {f:3}) is 6
PASS bar(1, 2, {f:3}) is 6
PASS bar(1, 2, {f:3}) is 6
PASS bar(1, 2, {f:3}) is 6
PASS bar(1, 2, {f:3}) is 6
PASS bar(1, 2, {f:3}) is 6
PASS bar(1, 2, {f:3}) is 6
PASS bar(1, 2, {f:3}) is 6
PASS bar(1, 2, {f:3}) is 6
PASS bar(1, 2, {f:3}) is 6
PASS bar(1, 2, {f:3}) is 6
PASS bar(1, 2, {f:3}) is 6
PASS bar(1, 2, {f:3}) is 6
PASS bar(1, 2, {f:3}) is 6
PASS bar(1, 2, {f:3}) is 6
PASS bar(1, 2, {f:3}) is 6
PASS bar(1, 2, {f:3}) is 6
PASS bar(1, 2, {f:3}) is 6
PASS bar(1, 2, {f:3}) is 6
PASS bar(1, 2, {f:3}) is 6
PASS bar(1, 2, {f:3}) is 6
PASS bar(1, 2, {f:3}) is 6
PASS bar(1, 2, {f:3}) is 6
PASS bar(1, 2, {f:3}) is 6
PASS bar(1, 2, {f:3}) is 6
PASS bar(1, 2, {f:3}) is 6
PASS bar(1, 2, {f:3}) is 6
PASS bar(1, 2, {f:3}) is 6
PASS bar(1, 2, {f:3}) is 6
PASS bar(1, 2, {f:3}) is 6
PASS bar(1, 2, {f:3}) is 6
PASS bar(1, 2, {f:3}) is 6
PASS bar(1, 2, {f:3}) is 6
PASS bar(1, 2, {f:3}) is 6
PASS bar(1, 2, {f:3}) is 6
PASS bar(1, 2, {f:3}) is 6
PASS bar(1, 2, {f:3}) is 6
PASS bar(1, 2, {f:3}) is 6
PASS bar(1, 2, {f:3}) is 6
PASS bar(1, 2, {f:3}) is 6
PASS bar(1, 2, {f:3}) is 6
PASS bar(1, 2, {f:3}) is 6
PASS bar(1, 2, {f:3}) is 6
PASS bar(1, 2, {f:3}) is 6
PASS bar(1, 2, {f:3}) is 6
PASS bar(1, 2, {f:3}) is 6
PASS bar(1, 2, {f:3}) is 6
PASS bar(1, 2, {f:3}) is 6
PASS bar(1, 2, {f:3}) is 6
PASS bar(1, 2, {f:3}) is 6
PASS bar(1, 2, {f:3}) is 6
PASS bar(1, 2, {f:3}) is 6
PASS bar(1, 2, {f:3}) is 6
PASS bar(1, 2, {f:3}) is 6
PASS bar(1, 2, {f:3}) is 6
PASS bar(2147483645, 2147483644, {f:9007199254740990}) is -8
PASS bar(2147483643, 2147483643, {f:18014398509481980}) is -16
PASS bar(2147483643, 2147483642, {f:36028797018963960}) is -16
PASS bar(2147483642, 2147483642, {f:36028797018963960}) is -16
PASS bar(2147483641, 2147483640, {f:144115188075855840}) is -32
PASS bar(2147483640, 2147483640, {f:144115188075855840}) is -64
PASS bar(2147483640, 2147483639, {f:288230376151711680}) is -64
PASS bar(2147483639, 2147483639, {f:288230376151711680}) is -64
PASS successfullyParsed is true

TEST COMPLETE
